Improves bitbang programming by supporting serials
Adds support for SERIAL environment variable for
bitbanging. The serial is piped to avrdude ( which currently needs
patch) as extended parameter.

Removes some duplicated lines in targets by converting the bang target
in a target which runs the other.

Adds LFUSE,HFUSE,EFUSE variable to get one place to set the fuses.